DB IDs are 64-bit

+++ Code/usecase/src/main/scala/com/beardedlogic/usecase/model/Data.scala
@@ -2,12 +2,12 @@ package com.beardedlogic.usecase.model
 
 import scala.slick.driver.PostgresDriver.simple._
 
-case class Data(id: Int,
+case class Data(id: Long,
                 data_type: DataType)
 
 object DataTable extends Table[Data]("data") {
   implicit val typeMapper = MappedTypeMapper.base[DataType, Short](_.ordinal, DataType(_))
-  def id = column[Int]("id")
+  def id = column[Long]("id")
   def data_type = column[DataType]("type_id")
   def * = id ~ data_type <>(Data, Data.unapply _)
 
@@ -16,6 +16,6 @@ object DataTable extends Table[Data]("data") {
     Data(newId, dataType)
   }
 
-  val QueryByID = for {id <- Parameters[Int]; r <- this if r.id is id} yield r
-  def apply(id: Int)(implicit s: Session): Data = QueryByID(id).first
+  val QueryByID = for {id <- Parameters[Long]; r <- this if r.id is id} yield r
+  def apply(id: Long)(implicit s: Session): Data = QueryByID(id).first
 }
